moved choice to util and tested
