+++ ledger/controllers/surveyor.js
@@ -8,6 +8,7 @@ const surveyors = require('../lib/surveyor')
 const utils = require('bat-utils')
 const braveHapi = utils.extras.hapi
 const braveJoi = utils.extras.joi
+const { surveyorChoices } = utils.extras.utils
 
 const defaultCohorts = ['control', 'grant', 'ads']
 const v1 = {}
@@ -763,26 +764,5 @@ async function addSurveyorChoices (runtime, surveyor = {}) {
 
 async function getAdjustedChoices (runtime, base, currencies) {
   const rates = await runtime.currency.rates(base, currencies)
-  return underscore.mapObject(rates, choicesPrices)
-}
-
-function choicesPrices (ratio) {
-  const table = [
-    [3, 5, 7, 10, 20],
-    [4, 6, 9, 12, 25],
-    [5, 8, 11, 17, 35],
-    [6, 10, 14, 20, 40],
-    [9, 12, 20, 35, 50],
-    [15, 25, 35, 50, 100],
-    [20, 35, 50, 85],
-    [30, 50, 70, 100]
-  ]
-  const priceIncrements = [1, 0.8, 0.6, 0.5, 0.35, 0.2, 0.15, 0.1]
-  let index = underscore.findIndex(priceIncrements, (increment) => {
-    return increment <= ratio
-  })
-  if (underscore.isUndefined(index)) {
-    index = priceIncrements.length - 1
-  }
-  return table[index]
+  return underscore.mapObject(rates, surveyorChoices)
 }
